Previo reactivar gitignore

+++ staticsite/sitebuilder/management/commands/build.py
@@ -3,7 +3,7 @@ import shutil
 
 from django.conf import settings
 from django.core.management import call_command
-from django.core.management.base import BaseCommand
+from django.core.management.base import BaseCommand, CommandError
 from django.core.urlresolvers import reverse
 from django.test.client import Client
 
@@ -17,20 +17,39 @@ class Command(BaseCommand):
 
     help  = 'Build static site outut'
 
+    def add_arguments(self, parser):
+        parser.add_argument('page', nargs='+', type=str)
+
     def handle(self, *args, **options):
         """Request pages and build output"""
+        print(options['page'][0])
+        if options['page'] is not None:
+            print('Estamos aqui')
+            pages = options['page']
+            available = list(get_pages())
+            invalid = []
 
-        if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
-            shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
-        
-        os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
-        os.makedirs(settings.STATIC_ROOT)
+            for page in pages:
+                if page not in available:
+                    invalid.append(page)
+
+            if invalid:
+                msg = 'Invalid pages: {} '.format(', '.join(invalid))
+                raise CommandError(msg)
+
+        else:
+            pages = get_pages()
+            if os.path.exists(settings.SITE_OUTPUT_DIRECTORY):
+                shutil.rmtree(settings.SITE_OUTPUT_DIRECTORY)
+            
+            os.mkdir(settings.SITE_OUTPUT_DIRECTORY)
+            os.makedirs(settings.STATIC_ROOT)
 
         call_command('collectstatic', interactive=False, clear=True, verbosity = 0)
 
         client = Client()
 
-        for page in get_pages():
+        for page in pages:
             url = reverse('page' , kwargs = {'slug': page})
             print(url)
 
@@ -40,7 +59,9 @@ class Command(BaseCommand):
                 output_dir = settings.SITE_OUTPUT_DIRECTORY
             else :
                 output_dir = os.path.join(settings.SITE_OUTPUT_DIRECTORY, page)
-                os.makedirs(output_dir)
+                if not os.path.exists(output_dir):
+                    os.makedirs(output_dir)
+              
             with open(os.path.join(output_dir, 'index.html') ,'wb') as f:
                 f.write(response.content)
